requirements.txt
torch
numpy
scikit-learn
sentence-transformers
requests
redis
pydantic
pydantic-settings
python-dotenv
joblib